requirements.txt
altgraph==0.17
anyio==3.6.2
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.5.7
backports.zoneinfo==0.2.1
black==21.7b0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==8.0.1
cloudinary==1.26.0
colorama==0.4.4
conda==4.8.3
conda-package-handling==1.7.0
contourpy==1.0.6
cryptography==2.9.2
cycler==0.11.0
Django==3.0.7
django-extensions==3.0.2
django-jazzmin==2.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
Eel==0.14.0
fastapi==0.92.0
flake8==3.9.2
Flask==2.2.3
fonttools==4.38.0
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
graphviz==0.14
greenlet==1.1.0
h11==0.14.0
httptools==0.5.0
idna==2.9
importlib-metadata==6.0.0
isort==5.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.2
mccabe==0.6.1
menuinst==1.4.16
mypy-extensions==0.4.3
mysql-connector-python==8.0.20
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pathspec==0.8.1
pefile==2021.5.24
Pillow==8.3.2
platformdirs==2.5.2
protobuf==3.12.2
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.20
pydantic==1.10.5
pydot==1.4.1
pyflakes==2.3.1
pygame==2.0.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==6.0
regex==2021.7.6
requests==2.23.0
ruamel_yaml==0.15.87
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.1
six==1.15.0
sklearn==0.0.post1
sniffio==1.3.0
sqlparse==0.3.1
starlette==0.25.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==1.0.4
tqdm==4.46.0
typing_extensions==4.3.0
tzdata==2022.1
unicorn==2.0.1.post1
urllib3==1.25.8
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.3
whichcraft==0.6.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.14.0
zope.event==4.5.0
zope.interface==5.4.0
